feat(api): Build repository and service layers

+++ api/brasileirao-api/src/main/java/com/selecaoglobocom/brasileirao_api/repository/TeamRepository.java
@@ -1,10 +1,9 @@
 package com.selecaoglobocom.brasileirao_api.repository;
 
-import com.selecaoglobocom.brasileirao_api.model.Team;
+import com.selecaoglobocom.brasileirao_api.entity.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeamRepository extends JpaRepository<Team, Long> {
-    // Já fornece métodos como findById(), findAll(), save(), delete(), etc.
+public interface TeamRepository extends JpaRepository<Team, Integer> {
 }
